LayoutTests/css3/flexbox/position-absolute-children.html

<!DOCTYPE html>
<html>
<style>
body {
    margin: 0;
}

.title {
    margin-top: 1em;
}

.flexbox {
    display: -webkit-flex;
    background-color: #aaa;
    position: relative;
}
.flexbox div {
    border: 0;
    -webkit-flex: none;
}

.horizontal-tb {
    -webkit-writing-mode: horizontal-tb;
}
.horizontal-bt {
    -webkit-writing-mode: horizontal-bt;
}
.vertical-rl {
    -webkit-writing-mode: vertical-rl;
}
.vertical-lr {
    -webkit-writing-mode: vertical-lr;
}

.row {
    -webkit-flex-flow: row;
}
.row-reverse {
    -webkit-flex-flow: row-reverse;
}
.column {
    -webkit-flex-flow: column;
}
.column-reverse {
    -webkit-flex-flow: column-reverse;
}

.flexbox :nth-child(1) {
    background-color: blue;
}
.flexbox :nth-child(2) {
    background-color: green;
}

.rtl {
    direction: rtl;
}
.ltr {
    direction: ltr;
}

.justify-content-flex-start {
    -webkit-justify-content: flex-start;
}
.justify-content-flex-end {
    -webkit-justify-content: flex-end;
}
.justify-content-center {
    -webkit-justify-content: center;
}
.justify-content-space-between {
    -webkit-justify-content: space-between;
}
.justify-content-space-around {
    -webkit-justify-content: space-around;
}
</style>
<script src="../../resources/check-layout.js"></script>
<body onload="checkLayout('.flexbox')">
<script>
var expectations = {
    'horizontal-tb': {
        'row': {
            'ltr': {
                'flex-start': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'flex-end': {
                    'child1': [80, 0],
                    'child2': [80, 0],
                },
                'center': {
                    'child1': [40, 0],
                    'child2': [40, 0],
                },
                'space-between': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'space-around': {
                    'child1': [40, 0],
                    'child2': [40, 0],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [40, 0],
                    'child2': [70, 0],
                },
                'flex-end': {
                    'child1': [-40, 0],
                    'child2': [-10, 0],
                },
                'center': {
                    'child1': [0, 0],
                    'child2': [30, 0],
                },
                'space-between': {
                    'child1': [40, 0],
                    'child2': [70, 0],
                },
                'space-around': {
                    'child1': [0, 0],
                    'child2': [30, 0],
                },
            },
        },
        'column': {
            'ltr': {
                'flex-start': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'flex-end': {
                    'child1': [0, 80],
                    'child2': [0, 80],
                },
                'center': {
                    'child1': [0, 40],
                    'child2': [0, 40],
                },
                'space-between': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'space-around': {
                    'child1': [0, 40],
                    'child2': [0, 40],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [10, 0],
                    'child2': [0, 0],
                },
                'flex-end': {
                    'child1': [10, 80],
                    'child2': [0, 80],
                },
                'center': {
                    'child1': [10, 40],
                    'child2': [0, 40],
                },
                'space-between': {
                    'child1': [10, 0],
                    'child2': [0, 0],
                },
                'space-around': {
                    'child1': [10, 40],
                    'child2': [0, 40],
                },
            },
        },
        'row-reverse': {
            'ltr': {
                'flex-start': {
                    'child1': [80, 0],
                    'child2': [80, 0],
                },
                'flex-end': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'center': {
                    'child1': [40, 0],
                    'child2': [40, 0],
                },
                'space-between': {
                    'child1': [80, 0],
                    'child2': [80, 0],
                },
                'space-around': {
                    'child1': [40, 0],
                    'child2': [40, 0],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [-40, 0],
                    'child2': [-10, 0],
                },
                'flex-end': {
                    'child1': [40, 0],
                    'child2': [70, 0],
                },
                'center': {
                    'child1': [0, 0],
                    'child2': [30, 0],
                },
                'space-between': {
                    'child1': [-40, 0],
                    'child2': [-10, 0],
                },
                'space-around': {
                    'child1': [0, 0],
                    'child2': [30, 0],
                },
            },
        },
        'column-reverse': {
            'ltr': {
                'flex-start': {
                    'child1': [0, 80],
                    'child2': [0, 80],
                },
                'flex-end': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'center': {
                    'child1': [0, 40],
                    'child2': [0, 40],
                },
                'space-between': {
                    'child1': [0, 80],
                    'child2': [0, 80],
                },
                'space-around': {
                    'child1': [0, 40],
                    'child2': [0, 40],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [10, 80],
                    'child2': [0, 80],
                },
                'flex-end': {
                    'child1': [10, 0],
                    'child2': [0, 0],
                },
                'center': {
                    'child1': [10, 40],
                    'child2': [0, 40],
                },
                'space-between': {
                    'child1': [10, 80],
                    'child2': [0, 80],
                },
                'space-around': {
                    'child1': [10, 40],
                    'child2': [0, 40],
                },
            },
        },
    },
    'horizontal-bt': {
        'row': {
            'ltr': {
                'flex-start': {
                    'child1': [0, 10],
                    'child2': [0, 0],
                },
                'flex-end': {
                    'child1': [80, 10],
                    'child2': [80, 0],
                },
                'center': {
                    'child1': [40, 10],
                    'child2': [40, 0],
                },
                'space-between': {
                    'child1': [0, 10],
                    'child2': [0, 0],
                },
                'space-around': {
                    'child1': [40, 10],
                    'child2': [40, 0],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [40, 10],
                    'child2': [70, 0],
                },
                'flex-end': {
                    'child1': [-40, 10],
                    'child2': [-10, 0],
                },
                'center': {
                    'child1': [0, 10],
                    'child2': [30, 0],
                },
                'space-between': {
                    'child1': [40, 10],
                    'child2': [70, 0],
                },
                'space-around': {
                    'child1': [0, 10],
                    'child2': [30, 0],
                },
            },
        },
        'column': {
            'ltr': {
                'flex-start': {
                    'child1': [0, 40],
                    'child2': [0, 70],
                },
                'flex-end': {
                    'child1': [0, -40],
                    'child2': [0, -10],
                },
                'center': {
                    'child1': [0, 0],
                    'child2': [0, 30],
                },
                'space-between': {
                    'child1': [0, 40],
                    'child2': [0, 70],
                },
                'space-around': {
                    'child1': [0, 0],
                    'child2': [0, 30],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [10, 40],
                    'child2': [0, 70],
                },
                'flex-end': {
                    'child1': [10, -40],
                    'child2': [0, -10],
                },
                'center': {
                    'child1': [10, 0],
                    'child2': [0, 30],
                },
                'space-between': {
                    'child1': [10, 40],
                    'child2': [0, 70],
                },
                'space-around': {
                    'child1': [10, 0],
                    'child2': [0, 30],
                },
            },
        },
        'row-reverse': {
            'ltr': {
                'flex-start': {
                    'child1': [80, 10],
                    'child2': [80, 0],
                },
                'flex-end': {
                    'child1': [0, 10],
                    'child2': [0, 0],
                },
                'center': {
                    'child1': [40, 10],
                    'child2': [40, 0],
                },
                'space-between': {
                    'child1': [80, 10],
                    'child2': [80, 0],
                },
                'space-around': {
                    'child1': [40, 10],
                    'child2': [40, 0],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [-40, 10],
                    'child2': [-10, 0],
                },
                'flex-end': {
                    'child1': [40, 10],
                    'child2': [70, 0],
                },
                'center': {
                    'child1': [0, 10],
                    'child2': [30, 0],
                },
                'space-between': {
                    'child1': [-40, 10],
                    'child2': [-10, 0],
                },
                'space-around': {
                    'child1': [0, 10],
                    'child2': [30, 0],
                },
            },
        },
        'column-reverse': {
            'ltr': {
                'flex-start': {
                    'child1': [0, -40],
                    'child2': [0, -10],
                },
                'flex-end': {
                    'child1': [0, 40],
                    'child2': [0, 70],
                },
                'center': {
                    'child1': [0, 0],
                    'child2': [0, 30],
                },
                'space-between': {
                    'child1': [0, -40],
                    'child2': [0, -10],
                },
                'space-around': {
                    'child1': [0, 0],
                    'child2': [0, 30],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [10, -40],
                    'child2': [0, -10],
                },
                'flex-end': {
                    'child1': [10, 40],
                    'child2': [0, 70],
                },
                'center': {
                    'child1': [10, 0],
                    'child2': [0, 30],
                },
                'space-between': {
                    'child1': [10, -40],
                    'child2': [0, -10],
                },
                'space-around': {
                    'child1': [10, 0],
                    'child2': [0, 30],
                },
            },
        },
    },
    'vertical-rl': {
        'row': {
            'ltr': {
                'flex-start': {
                    'child1': [10, 0],
                    'child2': [0, 0],
                },
                'flex-end': {
                    'child1': [10, 80],
                    'child2': [0, 80],
                },
                'center': {
                    'child1': [10, 40],
                    'child2': [0, 40],
                },
                'space-between': {
                    'child1': [10, 0],
                    'child2': [0, 0],
                },
                'space-around': {
                    'child1': [10, 40],
                    'child2': [0, 40],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [10, 40],
                    'child2': [0, 70],
                },
                'flex-end': {
                    'child1': [10, -40],
                    'child2': [0, -10],
                },
                'center': {
                    'child1': [10, 0],
                    'child2': [0, 30],
                },
                'space-between': {
                    'child1': [10, 40],
                    'child2': [0, 70],
                },
                'space-around': {
                    'child1': [10, 0],
                    'child2': [0, 30],
                },
            },
        },
        'column': {
            'ltr': {
                'flex-start': {
                    'child1': [40, 0],
                    'child2': [70, 0],
                },
                'flex-end': {
                    'child1': [-40, 0],
                    'child2': [-10, 0],
                },
                'center': {
                    'child1': [0, 0],
                    'child2': [30, 0],
                },
                'space-between': {
                    'child1': [40, 0],
                    'child2': [70, 0],
                },
                'space-around': {
                    'child1': [0, 0],
                    'child2': [30, 0],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [40, 10],
                    'child2': [70, 0],
                },
                'flex-end': {
                    'child1': [-40, 10],
                    'child2': [-10, 0],
                },
                'center': {
                    'child1': [0, 10],
                    'child2': [30, 0],
                },
                'space-between': {
                    'child1': [40, 10],
                    'child2': [70, 0],
                },
                'space-around': {
                    'child1': [0, 10],
                    'child2': [30, 0],
                },
            },
        },
        'row-reverse': {
            'ltr': {
                'flex-start': {
                    'child1': [10, 80],
                    'child2': [0, 80],
                },
                'flex-end': {
                    'child1': [10, 0],
                    'child2': [0, 0],
                },
                'center': {
                    'child1': [10, 40],
                    'child2': [0, 40],
                },
                'space-between': {
                    'child1': [10, 80],
                    'child2': [0, 80],
                },
                'space-around': {
                    'child1': [10, 40],
                    'child2': [0, 40],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [10, -40],
                    'child2': [0, -10],
                },
                'flex-end': {
                    'child1': [10, 40],
                    'child2': [0, 70],
                },
                'center': {
                    'child1': [10, 0],
                    'child2': [0, 30],
                },
                'space-between': {
                    'child1': [10, -40],
                    'child2': [0, -10],
                },
                'space-around': {
                    'child1': [10, 0],
                    'child2': [0, 30],
                },
            },
        },
        'column-reverse': {
            'ltr': {
                'flex-start': {
                    'child1': [-40, 0],
                    'child2': [-10, 0],
                },
                'flex-end': {
                    'child1': [40, 0],
                    'child2': [70, 0],
                },
                'center': {
                    'child1': [0, 0],
                    'child2': [30, 0],
                },
                'space-between': {
                    'child1': [-40, 0],
                    'child2': [-10, 0],
                },
                'space-around': {
                    'child1': [0, 0],
                    'child2': [30, 0],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [-40, 10],
                    'child2': [-10, 0],
                },
                'flex-end': {
                    'child1': [40, 10],
                    'child2': [70, 0],
                },
                'center': {
                    'child1': [0, 10],
                    'child2': [30, 0],
                },
                'space-between': {
                    'child1': [-40, 10],
                    'child2': [-10, 0],
                },
                'space-around': {
                    'child1': [0, 10],
                    'child2': [30, 0],
                },
            },
        },
    },
    'vertical-lr': {
        'row': {
            'ltr': {
                'flex-start': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'flex-end': {
                    'child1': [0, 80],
                    'child2': [0, 80],
                },
                'center': {
                    'child1': [0, 40],
                    'child2': [0, 40],
                },
                'space-between': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'space-around': {
                    'child1': [0, 40],
                    'child2': [0, 40],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [0, 40],
                    'child2': [0, 70],
                },
                'flex-end': {
                    'child1': [0, -40],
                    'child2': [0, -10],
                },
                'center': {
                    'child1': [0, 0],
                    'child2': [0, 30],
                },
                'space-between': {
                    'child1': [0, 40],
                    'child2': [0, 70],
                },
                'space-around': {
                    'child1': [0, 0],
                    'child2': [0, 30],
                },
            },
        },
        'column': {
            'ltr': {
                'flex-start': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'flex-end': {
                    'child1': [80, 0],
                    'child2': [80, 0],
                },
                'center': {
                    'child1': [40, 0],
                    'child2': [40, 0],
                },
                'space-between': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'space-around': {
                    'child1': [40, 0],
                    'child2': [40, 0],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [0, 10],
                    'child2': [0, 0],
                },
                'flex-end': {
                    'child1': [80, 10],
                    'child2': [80, 0],
                },
                'center': {
                    'child1': [40, 10],
                    'child2': [40, 0],
                },
                'space-between': {
                    'child1': [0, 10],
                    'child2': [0, 0],
                },
                'space-around': {
                    'child1': [40, 10],
                    'child2': [40, 0],
                },
            },
        },
        'row-reverse': {
            'ltr': {
                'flex-start': {
                    'child1': [0, 80],
                    'child2': [0, 80],
                },
                'flex-end': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'center': {
                    'child1': [0, 40],
                    'child2': [0, 40],
                },
                'space-between': {
                    'child1': [0, 80],
                    'child2': [0, 80],
                },
                'space-around': {
                    'child1': [0, 40],
                    'child2': [0, 40],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [0, -40],
                    'child2': [0, -10],
                },
                'flex-end': {
                    'child1': [0, 40],
                    'child2': [0, 70],
                },
                'center': {
                    'child1': [0, 0],
                    'child2': [0, 30],
                },
                'space-between': {
                    'child1': [0, -40],
                    'child2': [0, -10],
                },
                'space-around': {
                    'child1': [0, 0],
                    'child2': [0, 30],
                },
            },
        },
        'column-reverse': {
            'ltr': {
                'flex-start': {
                    'child1': [80, 0],
                    'child2': [80, 0],
                },
                'flex-end': {
                    'child1': [0, 0],
                    'child2': [0, 0],
                },
                'center': {
                    'child1': [40, 0],
                    'child2': [40, 0],
                },
                'space-between': {
                    'child1': [80, 0],
                    'child2': [80, 0],
                },
                'space-around': {
                    'child1': [40, 0],
                    'child2': [40, 0],
                },
            },
            'rtl': {
                'flex-start': {
                    'child1': [80, 10],
                    'child2': [80, 0],
                },
                'flex-end': {
                    'child1': [0, 10],
                    'child2': [0, 0],
                },
                'center': {
                    'child1': [40, 10],
                    'child2': [40, 0],
                },
                'space-between': {
                    'child1': [80, 10],
                    'child2': [80, 0],
                },
                'space-around': {
                    'child1': [40, 10],
                    'child2': [40, 0],
                },
            },
        },
    },
};

var writingModes = ['horizontal-tb', 'horizontal-bt', 'vertical-rl', 'vertical-lr'];
var flexDirections = ['row', 'column', 'row-reverse', 'column-reverse'];
var directions = ['ltr', 'rtl'];
var justifyContents = ['flex-start', 'flex-end', 'center', 'space-between', 'space-around'];

function mainAxisDirection(writingMode, flexDirection)
{
    if ((writingMode.indexOf('horizontal') != -1 && flexDirection.indexOf('row') != -1)
        || (writingMode.indexOf('vertical') != -1 && flexDirection.indexOf('column') != -1))
        return 'width';
    return 'height';
}

function addChild(flexbox, mainAxis, crossAxis, mainAxisLength, crossAxisLength, expectations)
{
    var child = document.createElement('div');
    child.setAttribute('style', mainAxis + ': ' + mainAxisLength + 'px;'
        + crossAxis + ': ' + crossAxisLength + 'px; position: absolute;');

    child.setAttribute("data-expected-" + mainAxis, mainAxisLength);
    child.setAttribute("data-expected-" + crossAxis, crossAxisLength);
    child.setAttribute("data-offset-x", expectations[0]);
    child.setAttribute("data-offset-y", expectations[1]);

    flexbox.appendChild(child);
}

writingModes.forEach(function(writingMode) {
    flexDirections.forEach(function(flexDirection) {
        directions.forEach(function(direction) {
            justifyContents.forEach(function(justifyContent) {
                var flexboxClassName = writingMode + ' ' + direction + ' ' + flexDirection + ' justify-content-' + justifyContent;
                var title = document.createElement('div');
                title.className = 'title';
                title.innerHTML = flexboxClassName;
                document.body.appendChild(title);

                var mainAxis = mainAxisDirection(writingMode, flexDirection);
                var crossAxis = (mainAxis == 'width') ? 'height' : 'width';

                var flexbox = document.createElement('div');
                flexbox.className = 'flexbox ' + flexboxClassName;
                flexbox.setAttribute('style', mainAxis + ': 80px;' + crossAxis + ': 20px');

                var baselineMargin = (flexDirection.indexOf('row') != -1) ? '-webkit-margin-before: 5px' : '-webkit-margin-start: 5px';

                var testExpectations = expectations[writingMode][flexDirection][direction][justifyContent];
                addChild(flexbox, mainAxis, crossAxis, 40, 10, testExpectations['child1']);
                addChild(flexbox, mainAxis, crossAxis, 10, 20, testExpectations['child2']);

                document.body.appendChild(flexbox);
            })
        })
    })
})

</script>

</body>
